Move message logging extension to its own file.
